refactored Symbol to use mixins
