+++ symai/backend/mixin/__init__.py
@@ -1,0 +1,18 @@
+from .symbol_primitives import *
+
+SYMBOL_PRIMITIVES = [
+    ContextualPrimitives,
+    ValueHandlingPrimitives,
+    ComparisonPrimitives,
+    ExpressionHandlingPrimitives,
+    DataHandlingPrimitives,
+    UniquenessPrimitives,
+    PatternMatchingPrimitives,
+    QueryHandlingPrimitives,
+    ExecutionControlPrimitives,
+    TemplateStylingPrimitives,
+    DataClusteringPrimitives,
+    PersistencePrimitives,
+    OutputHandlingPrimitives
+]
+
